+++ ladderly-io/src/app/job-search/job-post/[id]/JobPostDetails.tsx
@@ -24,16 +24,16 @@ import { Form, FORM_ERROR, type FormProps } from '~/app/core/components/Form'
 import LabeledTextField from '~/app/core/components/LabeledTextField'
 import LabeledSelectField from '~/app/core/components/LabeledSelectField'
 import LabeledCheckboxField from '~/app/core/components/LabeledCheckboxField'
-import { z } from 'zod'
 
 // --- Helper & Other Components ---
 import { JobStepsSection } from './JobStepsSection'
-// Import the new form component and its types
+// Import the new form components and their types
 import {
   EditJobPostForm,
   JobPostEditSchema,
   type JobPostEditValues,
 } from './EditJobPostForm'
+import { AddJobSearchStepForm } from './AddJobSearchStepForm' // Import the new step form
 
 // Helper to format date to YYYY-MM-DD
 const formatDateForInput = (date: Date | string | undefined | null): string => {
@@ -43,277 +43,6 @@ const formatDateForInput = (date: Date | string | undefined | null): string => {
   } catch (e) {
     return ''
   }
-}
-
-// Component to create a new job search step
-interface AddJobSearchStepProps {
-  jobPostId: number
-  currentStatus: JobApplicationStatus
-  onSuccess: () => void
-  onStatusChange: (newStatus: JobApplicationStatus) => void
-}
-
-const AddJobSearchStep: React.FC<AddJobSearchStepProps> = ({
-  jobPostId,
-  currentStatus,
-  onSuccess,
-  onStatusChange,
-}) => {
-  const [isOpen, setIsOpen] = useState(false)
-  const [stepKind, setStepKind] = useState<JobSearchStepKind>(
-    JobSearchStepKind.OTHER,
-  )
-  const [date, setDate] = useState<string>(
-    new Date().toISOString().substring(0, 10),
-  )
-  const [notes, setNotes] = useState('')
-  const [isPassed, setIsPassed] = useState<boolean | null>(null)
-  const [isInPerson, setIsInPerson] = useState(false)
-  const [isSubmitting, setIsSubmitting] = useState(false)
-
-  const { mutate: createJobSearchStep } =
-    api.jobSearch.createJobSearchStep.useMutation({
-      onSuccess: () => {
-        // Auto-update application status based on step type and result
-        updateStatusBasedOnStep(stepKind, isPassed)
-        onSuccess()
-        setIsOpen(false)
-        resetForm()
-      },
-      onError: (error) => {
-        console.error('Failed to create job search step:', error)
-        setIsSubmitting(false)
-      },
-    })
-
-  const resetForm = () => {
-    setStepKind(JobSearchStepKind.OTHER)
-    setDate(new Date().toISOString().substring(0, 10))
-    setNotes('')
-    setIsPassed(null)
-    setIsInPerson(false)
-    setIsSubmitting(false)
-  }
-
-  const updateStatusBasedOnStep = (
-    kind: JobSearchStepKind,
-    passed: boolean | null,
-  ) => {
-    // Logic to automatically update application status based on step type and result
-    let newStatus: JobApplicationStatus | null = null
-
-    // Define interview step kinds
-    const interviewStepKinds = [
-      JobSearchStepKind.BEHAVIORAL_INTERVIEW,
-      JobSearchStepKind.TECHNICAL_CODE_SCREEN_MANUAL,
-      JobSearchStepKind.TECHNICAL_CODE_SCREEN_AUTOMATED,
-      JobSearchStepKind.SYSTEM_DESIGN,
-      JobSearchStepKind.TECHNICAL_CONVERSATION,
-      JobSearchStepKind.HIRING_MANAGER_CALL,
-      JobSearchStepKind.PHONE_SCREEN,
-      JobSearchStepKind.MULTI_ROUND_MULTI_KIND,
-    ] as JobSearchStepKind[]
-
-    // Any failed step should update to REJECTED
-    if (passed === false) {
-      newStatus = JobApplicationStatus.REJECTED
-    }
-
-    // If we've added an outbound message step and were previously not in outreach
-    if (
-      kind === JobSearchStepKind.OUTBOUND_MESSAGE &&
-      currentStatus !== JobApplicationStatus.IN_OUTREACH &&
-      passed !== false // Don't update if the step failed
-    ) {
-      newStatus = JobApplicationStatus.IN_OUTREACH
-    }
-
-    // If we added an interview step and weren't in interview phase yet
-    if (
-      interviewStepKinds.includes(kind) &&
-      currentStatus !== JobApplicationStatus.IN_INTERVIEW &&
-      passed !== false // Don't update if the step failed
-    ) {
-      newStatus = JobApplicationStatus.IN_INTERVIEW
-    }
-
-    // If we're doing a background check, likely getting an offer
-    if (
-      kind === JobSearchStepKind.BACKGROUND_OR_REFERENCE_CHECK &&
-      currentStatus !== JobApplicationStatus.OFFER_RECEIVED &&
-      passed !== false // Don't update if the step failed
-    ) {
-      newStatus = JobApplicationStatus.OFFER_RECEIVED
-    }
-
-    // If there's a status change, call the parent handler
-    if (newStatus && newStatus !== currentStatus) {
-      onStatusChange(newStatus)
-    }
-  }
-
-  const handleSubmit = (e: React.FormEvent) => {
-    e.preventDefault()
-    setIsSubmitting(true)
-
-    createJobSearchStep({
-      jobPostId,
-      date: new Date(date),
-      kind: stepKind,
-      notes: notes ?? undefined,
-      isPassed: isPassed ?? undefined,
-      isInPerson,
-    })
-  }
-
-  if (!isOpen) {
-    return (
-      <button
-        onClick={() => setIsOpen(true)}
-        className="mt-4 w-full rounded-md border border-blue-500 px-4 py-2 text-blue-500 hover:bg-blue-50"
-      >
-        + Add Application Step
-      </button>
-    )
-  }
-
-  return (
-    <div className="mt-4 rounded-md border border-gray-200 p-4">
-      <div className="mb-4 flex items-center justify-between">
-        <h3 className="text-lg font-medium">Add Application Step</h3>
-        <button
-          onClick={() => setIsOpen(false)}
-          className="text-gray-500 hover:text-gray-700"
-        >
-          <svg
-            xmlns="http://www.w3.org/2000/svg"
-            width="20"
-            height="20"
-            viewBox="0 0 24 24"
-            fill="none"
-            stroke="currentColor"
-            strokeWidth="2"
-            strokeLinecap="round"
-            strokeLinejoin="round"
-          >
-            <line x1="18" y1="6" x2="6" y2="18"></line>
-            <line x1="6" y1="6" x2="18" y2="18"></line>
-          </svg>
-        </button>
-      </div>
-
-      <form onSubmit={handleSubmit}>
-        <div className="mb-4">
-          <label className="mb-1 block text-sm font-medium">Step Type</label>
-          <select
-            value={stepKind}
-            onChange={(e) => setStepKind(e.target.value as JobSearchStepKind)}
-            className="w-full rounded-md border border-gray-300 p-2"
-            required
-          >
-            {Object.values(JobSearchStepKind).map((kind) => (
-              <option key={kind} value={kind}>
-                {kind.replace(/_/g, ' ')}
-              </option>
-            ))}
-          </select>
-        </div>
-
-        <div className="mb-4">
-          <label className="mb-1 block text-sm font-medium">Date</label>
-          <input
-            type="date"
-            value={date}
-            onChange={(e) => setDate(e.target.value)}
-            className="w-full rounded-md border border-gray-300 p-2"
-            required
-          />
-        </div>
-
-        <div className="mb-4">
-          <label className="mb-1 block text-sm font-medium">Notes</label>
-          <textarea
-            value={notes}
-            onChange={(e) => setNotes(e.target.value)}
-            className="w-full rounded-md border border-gray-300 p-2"
-            rows={3}
-            placeholder="Add any notes about this step"
-          />
-        </div>
-
-        <div className="mb-4 flex flex-col space-y-4 sm:flex-row sm:items-center sm:space-x-4 sm:space-y-0">
-          <div>
-            <label className="mb-1 block text-sm font-medium">Result</label>
-            <div className="flex flex-wrap gap-2">
-              <button
-                type="button"
-                onClick={() => setIsPassed(true)}
-                className={`rounded-md px-3 py-1 ${
-                  isPassed === true
-                    ? 'bg-green-500 text-white'
-                    : 'bg-gray-100 text-gray-700'
-                }`}
-              >
-                Passed
-              </button>
-              <button
-                type="button"
-                onClick={() => setIsPassed(false)}
-                className={`rounded-md px-3 py-1 ${
-                  isPassed === false
-                    ? 'bg-red-500 text-white'
-                    : 'bg-gray-100 text-gray-700'
-                }`}
-              >
-                Failed
-              </button>
-              <button
-                type="button"
-                onClick={() => setIsPassed(null)}
-                className={`rounded-md px-3 py-1 ${
-                  isPassed === null
-                    ? 'bg-blue-500 text-white'
-                    : 'bg-gray-100 text-gray-700'
-                }`}
-              >
-                Pending
-              </button>
-            </div>
-          </div>
-
-          <div className="flex items-center">
-            <input
-              type="checkbox"
-              id="isInPerson"
-              checked={isInPerson}
-              onChange={(e) => setIsInPerson(e.target.checked)}
-              className="size-4 rounded border-gray-300"
-            />
-            <label htmlFor="isInPerson" className="ml-2 text-sm font-medium">
-              In Person
-            </label>
-          </div>
-        </div>
-
-        <div className="flex justify-end space-x-2">
-          <button
-            type="button"
-            onClick={() => setIsOpen(false)}
-            className="rounded-md border border-gray-300 px-4 py-2 hover:bg-gray-50"
-          >
-            Cancel
-          </button>
-          <button
-            type="submit"
-            disabled={isSubmitting}
-            className="rounded-md bg-blue-500 px-4 py-2 text-white hover:bg-blue-600 disabled:opacity-50"
-          >
-            {isSubmitting ? 'Adding...' : 'Add Step'}
-          </button>
-        </div>
-      </form>
-    </div>
-  )
 }
 
 // Status badge component
